# openapi.yaml
openapi: 3.1.0
info:
  title: UI-Driven API Key Management System
  version: 0.1.0
  description: Initial scaffold of endpoints for user-ui and admin-ui.
servers:
  - url: http://localhost:3020
    description: user-ui
  - url: http://localhost:4000
    description: admin-ui
paths:
...
  /register:
    get:
      summary: Registration form (protected)
      responses:
        '200':
          description: Registration HTML form
    post:
      summary: Submit registration (protected, CSRF required except in test)
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                purpose:
                  type: string
                  enum: [internal_tooling, third_party_integration, other]
                projectName:
                  type: string
                contactEmail:
                  type: string
              required: [purpose]
      responses:
        '302':
          description: Redirect to /dashboard
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /keys:
    get:
      summary: List current user's API keys (metadata only)
      responses:
        '200':
          description: Keys HTML page (no plaintext)
    post:
      summary: Issue a new API key (display-once)
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                usageDescription:
                  type: string
              required: [usageDescription]
      responses:
        '200':
          description: Display-once page containing the plaintext key
        '429':
          $ref: '#/components/responses/RateLimited'
  /keys/{id}/revoke:
    post:
      summary: Revoke an API key
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '302':
          description: Redirect to /keys
        '429':
          $ref: '#/components/responses/RateLimited'
components:
  headers:
    RateLimit-Limit:
      description: Maximum number of requests allowed in the current window.
      schema:
        type: integer
    RateLimit-Remaining:
      description: Remaining number of requests in the current window.
      schema:
        type: integer
    RateLimit-Reset:
      description: Time when the current rate limit window resets in UTC epoch seconds.
      schema:
        type: integer
    Retry-After:
      description: Seconds to wait before making a new request.
      schema:
        type: integer
  responses:
    RateLimited:
      description: Too Many Requests (rate limited)
      headers:
        RateLimit-Limit:
          $ref: '#/components/headers/RateLimit-Limit'
        RateLimit-Remaining:
          $ref: '#/components/headers/RateLimit-Remaining'
        RateLimit-Reset:
          $ref: '#/components/headers/RateLimit-Reset'
        Retry-After:
          $ref: '#/components/headers/Retry-After'